Website_image_captioning/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

opencv-python-headless
transformers
pandas
python-dotenv
openai
pillow
numpy<2.0
av
tqdm
torch==2.1.2+cpu
flask
# Core dependencies
werkzeug
torchvision==0.16.2+cpu
